Addes comments to all solutions
